requirements.txt
certifi==2020.11.8
chardet==3.0.4
cycler==0.10.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
urllib3==1.26.2
